requirements.txt
antlr4-python3-runtime
Brotli==1.1.0
python_snappy==0.7.1
numpy==1.26.4
rdflib==7.0.0
scikit_learn==1.5.0
tqdm==4.66.4
zstandard==0.22.0
